vchat2.2/backend/requirements.txt
openai==1.3.8
python-dotenv==1.0.0
numpy==1.24.3
elevenlabs>=2.0.0
pydub>=0.25.1
simpleaudio>=1.0.4
tiktoken==0.5.2
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Firebase
firebase-admin==6.2.0
google-cloud-firestore==2.13.1

# CORS middleware
python-multipart==0.0.6

# Additional utilities
python-dateutil==2.8.2

# Audio processing for Linux (PyAudio alternative)
# PyAudio는 Linux에서 문제가 될 수 있으므로 조건부로 처리